update experiment models and api
