+++ vsn-backend/domain/model/experiment_models_test.go
@@ -1,4 +1,4 @@
-package experiment
+package model
 
 import (
 	"testing"
